+++ sqlobject/main.py
@@ -693,9 +693,9 @@ def deprecated(message, level=1, stacklevel=2):
     if warnings_level is not None and warnings_level <= level:
         warnings.warn(message, DeprecationWarning, stacklevel=stacklevel)
 
-# if sys.version_info[:2] < (2, 6):
-#     deprecated("Support for Python 2.5 has been declared obsolete "
-#     "and will be removed in the next release of SQLObject")
+if sys.version_info[:2] < (2, 7):
+    deprecated("Support for Python 2.6 has been declared obsolete "
+               "and will be removed in the next release of SQLObject")
 
 
 def setDeprecationLevel(warning=1, exception=None):